requirements-dev.txt
flask
Flask-OAuthlib
flask_sqlalchemy
flask_migrate
github3.py
psycopg2